moved emote handling to local emote module
