+++ src/app/__init__.py
@@ -8,7 +8,7 @@ from flask_socketio import SocketIO
 from flask_httpauth import HTTPTokenAuth
 from flask import redirect, request
 
-from .emote_handling import Emotes
+from .emotes.emote_handling import Emotes
 from .user_limiter import UserLimiter
 from .obj import UserManager, get_default_user
 from utils.shell import Console, white, green2, red
